Microsoft Internet Explorer fixes, Part 4 Call polyfillFixes() from a few places, but have it set a global within the file, so that it only runs once.  Then we should be hardened against variations in the way webpac compiles the bundles. Lint cleanup. Moved polyfillFixes() to cordovaUtils.js to avoid a circular dependency.
For wevote/WebApp/issues/3184
